requirements.txt
django==5.1.*
djangorestframework==3.15.*
django-environ==0.11.*
psycopg2-binary==2.9.*
sqlalchemy==2.0.*
pydantic-settings==2.6.*
gunicorn==22.0.*
google-cloud-aiplatform==1.71.*           # Vertex AI
langchain==0.3.*
langchain-community==0.3.*
langchain-google-vertexai==2.0.*
